+++ src/VersionControl/IGitRepository.cs
@@ -5,4 +5,5 @@ public interface IGitRepository : IDisposable
     void ForceCheckout(string branch);
     void Fetch(string remote = "origin");
     void ResetHard(string toBranch);
+    void SetRemote(string name, string newUrl);
 }